create export_exec_utils.c
